I think the variable name might have changed here
I don't think that somxzint1 is a valid name for MLD now. I've put in the names of two alternatives I found in recent runs of the model. However, I'm uncertain of the formatting of the statement for a yaml file. I've essentially assumed that it'll look along the list until it finds one that it likes. It probably doesn't work that way, however. So don't just accept this change without checking it's not stupid! :-)